+++ src/Atoms/Icon/action/copy/index.ts
@@ -1,5 +1,7 @@
+import Copy12 from './Copy12';
 import Copy16 from './Copy16';
 
 export default {
+  Copy12,
   Copy16,
 };